Remove database statement parser / analyzer deprecated code
