--- fe/fe-core/src/main/java/com/starrocks/analysis/AlterDatabaseRename.java
@@ -1,66 +1,0 @@
-// This file is made available under Elastic License 2.0.
-// This file is based on code available under the Apache license here:
-//   https://github.com/apache/incubator-doris/blob/master/fe/fe-core/src/main/java/org/apache/doris/analysis/AlterDatabaseRename.java
-
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package com.starrocks.analysis;
-
-import com.starrocks.sql.ast.AstVisitor;
-
-public class AlterDatabaseRename extends DdlStmt {
-    private String dbName;
-    private String newDbName;
-
-    public AlterDatabaseRename(String dbName, String newDbName) {
-        this.dbName = dbName;
-        this.newDbName = newDbName;
-    }
-
-    public String getDbName() {
-        return dbName;
-    }
-
-    public String getNewDbName() {
-        return newDbName;
-    }
-
-    public void setDbName(String dbName) {
-        this.dbName = dbName;
-    }
-
-    public void setNewDbName(String newName) {
-        this.newDbName = newName;
-    }
-
-    @Override
-    public String toSql() {
-        return "ALTER DATABASE " + dbName + " RENAME " + newDbName;
-    }
-
-    @Override
-    public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
-        return visitor.visitAlterDatabaseRename(this, context);
-    }
-
-    @Override
-    public boolean isSupportNewPlanner() {
-        return true;
-    }
-
-}
