+++ fe/fe-core/src/main/java/com/starrocks/execution/CreateDbExecutor.java
@@ -2,7 +2,6 @@
 
 package com.starrocks.execution;
 
-import com.starrocks.analysis.CreateDbStmt;
 import com.starrocks.analysis.StatementBase;
 import com.starrocks.common.AlreadyExistsException;
 import com.starrocks.common.DdlException;
@@ -10,6 +9,7 @@ import com.starrocks.common.ErrorCode;
 import com.starrocks.common.ErrorReport;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.qe.ShowResultSet;
+import com.starrocks.sql.ast.CreateDbStmt;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
