+++ fe/fe-core/src/main/java/com/starrocks/sql/ast/AlterDatabaseQuotaStmt.java
@@ -1,0 +1,58 @@
+// This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
+
+package com.starrocks.sql.ast;
+
+import com.starrocks.analysis.DdlStmt;
+
+public class AlterDatabaseQuotaStmt extends DdlStmt {
+    private String dbName;
+    private final QuotaType quotaType;
+    private final String quotaValue;
+    private long quota;
+
+    public enum QuotaType {
+        NONE,
+        DATA,
+        REPLICA
+    }
+
+    public AlterDatabaseQuotaStmt(String dbName, QuotaType quotaType, String quotaValue) {
+        this.dbName = dbName;
+        this.quotaType = quotaType;
+        this.quotaValue = quotaValue;
+    }
+
+    public String getDbName() {
+        return dbName;
+    }
+
+    public long getQuota() {
+        return quota;
+    }
+
+    public String getQuotaValue() {
+        return quotaValue;
+    }
+
+    public void setDbName(String dbName) {
+        this.dbName = dbName;
+    }
+
+    public void setQuota(long quota) {
+        this.quota = quota;
+    }
+
+    public QuotaType getQuotaType() {
+        return quotaType;
+    }
+
+    @Override
+    public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
+        return visitor.visitAlterDbQuotaStmt(this, context);
+    }
+
+    @Override
+    public boolean isSupportNewPlanner() {
+        return true;
+    }
+}
