+++ fe/fe-core/src/main/java/com/starrocks/sql/ast/AlterDatabaseRename.java
@@ -1,0 +1,39 @@
+// This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
+
+package com.starrocks.sql.ast;
+
+import com.starrocks.analysis.DdlStmt;
+import com.starrocks.sql.ast.AstVisitor;
+
+public class AlterDatabaseRename extends DdlStmt {
+    private String dbName;
+    private final String newDbName;
+
+    public AlterDatabaseRename(String dbName, String newDbName) {
+        this.dbName = dbName;
+        this.newDbName = newDbName;
+    }
+
+    public String getDbName() {
+        return dbName;
+    }
+
+    public String getNewDbName() {
+        return newDbName;
+    }
+
+    public void setDbName(String dbName) {
+        this.dbName = dbName;
+    }
+
+    @Override
+    public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
+        return visitor.visitAlterDatabaseRename(this, context);
+    }
+
+    @Override
+    public boolean isSupportNewPlanner() {
+        return true;
+    }
+
+}
