+++ fe/fe-core/src/main/java/com/starrocks/sql/ast/RecoverDbStmt.java
@@ -1,0 +1,31 @@
+// This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
+
+package com.starrocks.sql.ast;
+
+import com.starrocks.analysis.DdlStmt;
+
+public class RecoverDbStmt extends DdlStmt {
+    private String dbName;
+
+    public RecoverDbStmt(String dbName) {
+        this.dbName = dbName;
+    }
+
+    public String getDbName() {
+        return dbName;
+    }
+
+    public void setDbName(String dbname) {
+        this.dbName = dbname;
+    }
+
+    @Override
+    public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
+        return visitor.visitRecoverDbStmt(this, context);
+    }
+
+    @Override
+    public boolean isSupportNewPlanner() {
+        return true;
+    }
+}
