+++ fe/fe-core/src/main/java/com/starrocks/sql/ast/UseDbStmt.java
@@ -4,8 +4,6 @@ package com.starrocks.sql.ast;
 
 import com.starrocks.analysis.RedirectStatus;
 import com.starrocks.analysis.StatementBase;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 /**
  * Representation of a USE [catalog.]db statement.
@@ -13,13 +11,8 @@ import org.apache.logging.log4j.Logger;
  * Queries from JDBC will be handled by COM_QUERY protocol, it will generate UseDbStmt.
  */
 public class UseDbStmt extends StatementBase {
-    private static final Logger LOG = LogManager.getLogger(UseDbStmt.class);
     private final String catalog;
     private final String database;
-
-    public UseDbStmt(String db) {
-        this(null, db);
-    }
 
     public UseDbStmt(String catalog, String database) {
         this.catalog = catalog;
