+++ fe/fe-core/src/test/java/com/starrocks/analysis/DropMaterializedViewStmtNewPlannerTest.java
@@ -23,6 +23,7 @@ package com.starrocks.analysis;
 
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.server.GlobalStateMgr;
+import com.starrocks.sql.ast.CreateDbStmt;
 import com.starrocks.utframe.UtFrameUtils;
 import org.junit.AfterClass;
 import org.junit.Assert;
