+++ fe/fe-core/src/test/java/com/starrocks/catalog/CreateTableTest.java
@@ -22,7 +22,7 @@
 package com.starrocks.catalog;
 
 import com.starrocks.analysis.AlterTableStmt;
-import com.starrocks.analysis.CreateDbStmt;
+import com.starrocks.sql.ast.CreateDbStmt;
 import com.starrocks.analysis.CreateTableStmt;
 import com.starrocks.common.AnalysisException;
 import com.starrocks.common.Config;
