+++ fe/fe-core/src/test/java/com/starrocks/catalog/DropPartitionTest.java
@@ -25,7 +25,7 @@ import com.google.common.collect.Lists;
 import com.staros.proto.ObjectStorageInfo;
 import com.staros.proto.ShardStorageInfo;
 import com.starrocks.analysis.AlterTableStmt;
-import com.starrocks.analysis.CreateDbStmt;
+import com.starrocks.sql.ast.CreateDbStmt;
 import com.starrocks.analysis.CreateTableStmt;
 import com.starrocks.analysis.RecoverPartitionStmt;
 import com.starrocks.common.Config;
